+++ lib/src/controllers/vimeo_player_controller.dart
@@ -11,7 +11,7 @@ class VimeoPlayerController extends ValueNotifier<VimeoPlayerValue> {
   final String securityId;
   final String appId;
   final PlayerDeviceType type;
-  static ValueChanged<bool>? _controllerStateCallback;
+  static VoidCallback? _controllerStateCallback;
   VimeoPlayerController({
     required this.initialVideoId,
     this.flags = const VimeoPlayerFlags(),
@@ -19,8 +19,8 @@ class VimeoPlayerController extends ValueNotifier<VimeoPlayerValue> {
     required this.appId,
     required this.type,
   }) : super(VimeoPlayerValue(webViewController: null));
-  ValueChanged<bool>? get controllerStateCallback => _controllerStateCallback;
-  set controllerStateCallback(Function(bool)? b) {
+  VoidCallback? get controllerStateCallback => _controllerStateCallback;
+  set controllerStateCallback(VoidCallback? b) {
     _controllerStateCallback = b;
   }
 
@@ -34,7 +34,7 @@ class VimeoPlayerController extends ValueNotifier<VimeoPlayerValue> {
       updateValue(value.copyWith(isFullscreen: true));
 
   void reload() => value.webViewController?.reload();
-
+  void initialize() => _callMethod('initialize()');
   void play() => _callMethod('play()');
   void pause() => _callMethod('pause()');
   void mute() => _callMethod('mute()');
@@ -44,11 +44,13 @@ class VimeoPlayerController extends ValueNotifier<VimeoPlayerValue> {
   // bool get isFullscreen => _isFullscreen;
   // set isFullscreen(bool t) => _isFullscreen = t;
   _callMethod(String methodString) {
-    if (controllerStateCallback != null) {
-      controllerStateCallback!(value.isReady);
-    }
+    // if (controllerStateCallback != null) {
+    //   controllerStateCallback!(value.isReady);
+    // }
     if (value.isReady) {
       value.webViewController?.evaluateJavascript(source: methodString);
+      if (controllerStateCallback == null) return;
+      controllerStateCallback!();
     } else {
       print('The controller is not ready for method calls.');
       throw ErrorDescription("The controller is not ready for method calls.");