+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.3.7
 * Added controller state callback
+    - added initialize on controller
 
 ## 1.3.6
 * Additional Settings/Params